projeto-root/requirements.txt
Flask
openai
twilio
langdetect
pandas
odfpy
python-dotenv